test(AccountControllerTest): refactor tests to support BigDecimal balance type
- Updated AccountControllerTest to handle Account balance as BigDecimal instead of float.
- Refactored test setup and assertions to work with BigDecimal values for balance in account creation and retrieval endpoints.
- Ensured that the tests for account retrieval, creation, and exception handling are consistent with the new BigDecimal type for balance.

+++ src/test/java/com/ng_billing/bank_management/controller/AccountControllerTest.java
@@ -17,6 +17,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import static org.mockito.Mockito.*;
@@ -46,7 +47,7 @@ class AccountControllerTest {
         MockitoAnnotations.openMocks(this);
 
         objectMapper = new ObjectMapper();
-        account = new Account(234, 170.07f);
+        account = new Account(234, BigDecimal.valueOf(170.07f));
         accountDTO = new AccountDTO(account.getAccountNumber(), account.getBalance());
     }
